implemented livestream fucntionality
